Unify LLM runtime for rollout capture
