+++ openspec/changes/feat-rl/tasks.md
@@ -6,8 +6,8 @@
 - [x] Add request, ack, cancel, and publisher event APIs.
 - [x] Add `RLEventSink` as an `EventStreamer` sink.
 - [x] Make rollout storage optional with `persist_rollout`.
-- [x] Move rollout LLM implementation to `cube_harness.rl.llm`.
-- [x] Keep `cube_harness.llm` as the shared/benchmark LLM layer.
+- [x] Keep the LLM runtime unified in `cube_harness.llm`.
+- [x] Keep rollout-only LLM configuration in `cube_harness.rl.llm`.
 - [x] Add RL recipes.
 - [x] Add rollout service tests and throughput opt-in tests.
 - [x] Move deterministic RL smoke to `scripts/smoke/rl_mock_multiturn_service.py`.
