+++ src/cube_harness/agents/react.py
@@ -4,12 +4,11 @@ from typing import TYPE_CHECKING
 from cube.core import Action, ActionSchema, Observation
 from cube.task import STOP_ACTION
 from litellm import Message
-from pydantic import SerializeAsAny
 from termcolor import colored
 
 from cube_harness.agent import Agent, AgentConfig, apply_description_overrides
 from cube_harness.core import AgentOutput
-from cube_harness.llm import BaseLLMConfig, Prompt
+from cube_harness.llm import LLMConfig, Prompt
 from cube_harness.utils import parse_actions
 
 if TYPE_CHECKING:
@@ -19,7 +18,7 @@ logger = logging.getLogger(__name__)
 
 
 class ReactAgentConfig(AgentConfig):
-    llm_config: SerializeAsAny[BaseLLMConfig]
+    llm_config: LLMConfig
     can_finish: bool = True
     max_actions: int = 10
     max_obs_chars: int = 100000  # truncate long observations to M chars
